refactor : storageService 코드 리팩토링

+++ src/main/java/com/nhnacademy/inventory/medicines/enviroment/service/MedicineEnvironmentService.java
@@ -1,7 +1,9 @@
 package com.nhnacademy.inventory.medicines.enviroment.service;
 
+import com.nhnacademy.inventory.global.exception.ForbiddenException;
 import com.nhnacademy.inventory.global.util.UserContext;
 
+import com.nhnacademy.inventory.medicines.enviroment.domain.EnvironmentType;
 import com.nhnacademy.inventory.medicines.enviroment.domain.MedicineEnvironmentStandard;
 
 import com.nhnacademy.inventory.medicines.enviroment.domain.MedicineEnvironmentType;
@@ -10,13 +12,13 @@ import com.nhnacademy.inventory.medicines.enviroment.exception.EnvironmentStanda
 import com.nhnacademy.inventory.medicines.enviroment.operation.EnvironmentStandardCreator;
 import com.nhnacademy.inventory.medicines.enviroment.repository.MedicineEnvironmentStandardRepository;
 
-
 import com.nhnacademy.inventory.medicines.enviroment.repository.MedicineEnvironmentTypeRepository;
-import com.nhnacademy.inventory.medicines.medicine.exception.MedicineNotFoundException;
 import com.nhnacademy.inventory.organizations.member.domain.OrganizationMember;
+import com.nhnacademy.inventory.organizations.member.domain.OrganizationRole;
 import com.nhnacademy.inventory.organizations.member.repository.OrganizationMemberRepository;
 import com.nhnacademy.inventory.organizations.organization.domain.Organization;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 import com.nhnacademy.inventory.organizations.organization.exception.UserOrgNotFoundException;
@@ -36,13 +38,19 @@ public class MedicineEnvironmentService {
     private final EnvironmentStandardCreator environmentStandardCreator;
 
 
-    // 환경유형 저장 및 수정
+    // 입고용 환경기준, 유형 저장
     @Transactional
     public void createTypes(Long medicinePackageUnitId ,MedicineEnvironmentRequest request){
 
         UUID accountId = UserContext.getUserUuid();
         OrganizationMember organizationMember = memberRepository.findByAccountUuid(accountId)
                 .orElseThrow(UserOrgNotFoundException::new);
+
+
+        if(organizationMember.getOrganizationRole() == OrganizationRole.ORG_MEMBER){
+            throw new ForbiddenException();
+        }
+
         Organization organization = organizationMember.getOrganization();
         MedicineEnvironmentStandard standard = standardRepository.findByOrganizationIdAndPackageUnitIdForUpdate(organization.getId(),medicinePackageUnitId)
                 .orElse(null);
@@ -50,7 +58,6 @@ public class MedicineEnvironmentService {
 
         // 기준이 있다 그럼 저장 x 사용자는 조회된 기준으로 사용
         // -> 환경 기준 수정 부분에서 수정가능
-
 
         if(standard != null){
             return;
@@ -63,56 +70,90 @@ public class MedicineEnvironmentService {
 
     // request 0 -> 수정  request 모두 null -> 타입, 기준 모두 삭제
     @Transactional
-    public void updateTypes(Long medicinePackageUnitId,MedicineEnvironmentRequest request){
+    public void updateTypes(Long medicinePackageUnitId, MedicineEnvironmentRequest request){
 
         UUID accountId = UserContext.getUserUuid();
         OrganizationMember organizationMember = memberRepository.findByAccountUuid(accountId)
                 .orElseThrow(UserOrgNotFoundException::new);
+
+        if(organizationMember.getOrganizationRole() == OrganizationRole.ORG_MEMBER){
+            throw new ForbiddenException();
+        }
+
+
         Organization organization = organizationMember.getOrganization();
         MedicineEnvironmentStandard standard = standardRepository.findByOrganizationIdAndPackageUnitIdForUpdate(organization.getId(),medicinePackageUnitId)
-                .orElse(null);
+                .orElseThrow(EnvironmentStandardNotFoundException::new);
 
-        if(standard == null){
-            // 기준 추가
-            // 로직
-            environmentStandardCreator.createStandard(medicinePackageUnitId,organization,accountId,request);
-            return;
-        }
 
-        //기준이 있다.
-        typeRepository.deleteAllByMedicineEnvironmentStandardId(standard.getId());
-        List<MedicineEnvironmentType> types = environmentStandardCreator.makeTypes(standard,request);
+        MedicineEnvironmentType tepType = typeRepository.findByMedicineEnvironmentStandardAndEnvironmentType(standard,EnvironmentType.TEMPERATURE)
+                        .orElse(null);
+        MedicineEnvironmentType humType = typeRepository.findByMedicineEnvironmentStandardAndEnvironmentType(standard,EnvironmentType.HUMIDITY)
+            .orElse(null);
+        MedicineEnvironmentType illType = typeRepository.findByMedicineEnvironmentStandardAndEnvironmentType(standard,EnvironmentType.ILLUMINANCE)
+            .orElse(null);
 
-        if(types.isEmpty()){
-            standardRepository.deleteById(standard.getId());
-            return;
-        }
 
-        typeRepository.saveAll(types);
-        standard.updateIdAndAt(UserContext.getUserUuid());
+
+        updateEnvironment(standard,tepType,EnvironmentType.TEMPERATURE,request.minTemperature(),request.maxTemperature());
+        updateEnvironment(standard,humType, EnvironmentType.HUMIDITY,request.minHumidity(),request.maxHumidity());
+        updateEnvironment(standard,illType,EnvironmentType.ILLUMINANCE,request.minIlluminance(),request.maxIlluminance());
+
+        standard.updateIdAndAt(accountId);
 
     }
 
-
     // 환경 유형 삭제
     @Transactional
-    public void deleteTypes(Long standardId){
+    public void deleteTypes(Long packUnitId){
 
-        MedicineEnvironmentStandard standard = standardRepository.findById(standardId)
+        UUID accountId = UserContext.getUserUuid();
+        OrganizationMember organizationMember = memberRepository.findByAccountUuid(accountId)
+                .orElseThrow(UserOrgNotFoundException::new);
+
+        if(organizationMember.getOrganizationRole() == OrganizationRole.ORG_MEMBER){
+            throw new ForbiddenException();
+        }
+
+
+        MedicineEnvironmentStandard standard = standardRepository.findByMedicinePackageUnitId(packUnitId)
                         .orElseThrow(EnvironmentStandardNotFoundException::new);
 
         typeRepository.deleteAllByMedicineEnvironmentStandardId(standard.getId());
         standardRepository.deleteById(standard.getId());
 
     }
+    
 
+    private void updateEnvironment(MedicineEnvironmentStandard standard, MedicineEnvironmentType type, EnvironmentType status, BigDecimal min, BigDecimal max){
 
+        if(type == null){
 
+            if(min != null && max!= null){
 
+                MedicineEnvironmentType createType = MedicineEnvironmentType.create(standard,status,min,max);
+                typeRepository.save(createType);
 
+            }
 
+        }
+        else{
 
+            if(min == null && max == null){
+                typeRepository.deleteByMedicineEnvironmentStandardAndEnvironmentType(
+                        standard,status
+                );
 
+            }
+
+            else if(min != null && max != null){
+                type.update(min,max);
+
+            }
+
+        }
+
+    }
